find outliers

This function finds outliers in a dataset using quantile random forests.

Usage

outqrf(
  data,
  quantiles_type = 1000,
  threshold = 0.025,
  impute = TRUE,
  verbose = 1,
  weight = FALSE,
  ...
)

Arguments

data

a data frame

quantiles_type

'1000':seq(from = 0.001, to = 0.999, by = 0.001), '400':seq(0.0025,0.9975,0.0025)

threshold

a threshold for outlier detection

impute

a boolean value indicating whether to impute missing values

verbose

a boolean value indicating whether to print verbose output

weight

a boolean value indicating whether to use weight. if TRUE, The actual threshold will be threshold*r2.

...

additional arguments passed to the ranger function

Value

An object of class "outqrf" and a list with the following elements.

Data: Original data set in unchanged row order

outliers: Compact representation of outliers. Each row corresponds to an outlier and contains the following columns:

row: Row number of the outlier

col: Variable name of the outlier

observed: value of the outlier

predicted: predicted value of the outlier

rank: Rank of the outlier

outMatrix: Predicted value at different quantiles for each observation

r.squared: R-squared value of the quantile random forest model

outMatrix: Predicted value at different quantiles for each observation

r.squared: R-squared value of the quantile random forest model

oob.error: Out-of-bag error of the quantile random forest model

rmse: RMSE of the quantile random forest model

threshold: Threshold for outlier detection

Examples

iris_with_outliers <- generateOutliers(iris, p=0.05)
qrf = outqrf(iris_with_outliers)
#> 
#> Outlier identification by quantiles random forests
#> 
#>   Variables to check:		Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
#>   Variables used to check:	Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, Species
#> 
#>   Checking: Sepal.Length  Sepal.Width  Petal.Length  Petal.Width  
qrf$outliers
#>    row          col observed predicted  rank
#> 1   23 Sepal.Length     15.6      7.20 0.999
#> 2   89 Sepal.Length     -2.5      5.70 0.007
#> 3   98 Sepal.Length     -5.2      4.90 0.001
#> 4  133 Sepal.Length     -0.4      5.80 0.001
#> 5   30  Sepal.Width     10.6      3.40 0.999
#> 6   45  Sepal.Width      9.2      4.40 0.997
#> 7  140  Sepal.Width     -4.8      3.00 0.001
#> 8   77 Petal.Length     15.6      6.10 0.999
#> 9  112 Petal.Length     -3.7      5.00 0.003
#> 10 130 Petal.Length    -10.6      5.00 0.001
#> 11  28  Petal.Width      4.1      0.40 0.999
#> 12 129  Petal.Width     -5.3      1.55 0.001
evaluateOutliers(iris,iris_with_outliers,qrf$outliers)
#>     Actual  Predicted      Cover   Coverage Efficiency 
#>      31.00      12.00      12.00       0.39       1.00